+++ ui/packages/shared/profile/src/ProfileIcicleGraph/IcicleGraphArrow/ColorStackLegend.tsx
@@ -17,23 +17,28 @@ import {Icon} from '@iconify/react';
 import cx from 'classnames';
 
 import {useURLState} from '@parca/components';
-import {USER_PREFERENCES, useUserPreference} from '@parca/hooks';
-import {EVERYTHING_ELSE} from '@parca/store';
-import type {NavigateFunction} from '@parca/utilities';
+import {USER_PREFERENCES, useCurrentColorProfile, useUserPreference} from '@parca/hooks';
+import {EVERYTHING_ELSE, selectDarkMode, useAppSelector} from '@parca/store';
+import {type NavigateFunction} from '@parca/utilities';
 
-import {mappingColors} from './IcicleGraphNodes';
+import {getMappingColors} from '.';
+import useMappingList from './useMappingList';
 
 interface Props {
-  mappingColors: mappingColors;
+  mappings?: string[];
+  loading?: boolean;
   navigateTo?: NavigateFunction;
   compareMode?: boolean;
 }
 
 const ColorStackLegend = ({
-  mappingColors,
+  mappings,
   navigateTo,
   compareMode = false,
+  loading,
 }: Props): React.JSX.Element => {
+  const isDarkMode = useAppSelector(selectDarkMode);
+  const currentColorProfile = useCurrentColorProfile();
   const [colorProfileName] = useUserPreference<string>(
     USER_PREFERENCES.FLAMEGRAPH_COLOR_PROFILE.key
   );
@@ -41,6 +46,13 @@ const ColorStackLegend = ({
     param: 'binary_frame_filter',
     navigateTo,
   });
+
+  const mappingsList = useMappingList(mappings);
+
+  const mappingColors = useMemo(() => {
+    const colors = getMappingColors(mappingsList, isDarkMode, currentColorProfile);
+    return colors;
+  }, [isDarkMode, mappingsList, currentColorProfile]);
 
   const stackColorArray = useMemo(() => {
     return Object.entries(mappingColors).sort(([featureA], [featureB]) => {
@@ -54,7 +66,7 @@ const ColorStackLegend = ({
     });
   }, [mappingColors]);
 
-  if (mappingColors === undefined) {
+  if (stackColorArray.length === 0 && loading === false) {
     return <></>;
   }